pdf: fix drawquad
The preceding code resulted in errors making the
circles and ellipses look odd, compared to the
ps's ones.

The magic value b had to be replaced by a double,
for precision purposes.

Large dpis cause overflow on `ch*cv`. To address
this problem, we multiply the signs instead of
the numbers themselves.