pdf&ps: Add support for line control
This includes line thickness, line cap and line
join. This was added to ps using \X'set' commands.

A new drawing primitive, mirroring groff's, was
introduced: \D't'. It allows directly setting line
width using units. \D't 1i' has expected results
(a line width of an inch), whereas

.nr a 1i
.dv set linewidth \na

does not, because the argument of linewidth is in
thousandths of ems. It's also arguably more
cumbersome.