pdf: standalone pdfpos, replace longs by doubles
ints are 32-bit. Some users may use very high
dpis (eg, 720000). pdfpos00(x*100, y*100) makes
for frequent overflows, since x, y and dev_res
can be very large.

Doubles in pdfpos and pdfunit calculations:

This is very important when precision is required
(eg, when using grap). When ints are used instead,
the output is very choppy since lines' ends fall
only on points with integer coordinates (this
problem doesn't arise in ps because coordinates
are not divided by dev_res, so integer coordinates
are almost like floats because they are so dense).

This is also very problematic for tbl.